Linkedin Code
An almost working set that uses the Linkedin API. The issue is that when
the phone opens the web brower and authenticates on the Linkedin page
the brower is supposed to exit and return the user back to the
application. It doesn't do it on my phone. This step is crucial as it
delivers the access token needed to updates statues, search for people,
etc.