Scaled bitmap so no memory exceptions, new problem
Trying not to get memory errors when working with bitmaps.  Scaling the
image by 4.  This is still probably too much, so we might need to try x
8.

New problem: It seems that tesseract is not being able to initialize
correctly.  We might have to store stuff on the SD card instead of phone
internal memory.