Camera, fixed database bug, revert to API 8, execution time
I added some stuff to the camera code, and it no longer crashes for me.
I also fixed a bug in the database code that was incorrectly storing the
email address.
I also added in code that shows the execution time of the program
measured from the time the picture is taken to the time the results are
stored in the database.  It seems parsing the text takes the longest
time- up to 6 seconds if it manages to find a name, web address, and
phone number.  If no info is found, it takes about 2 seconds.

Note: LinkedIn will not work since I reverted the API back to 8 for
testing.