AsyncTask, Updated Linkedin Interface
I have tried to fix the AsyncTask to allow for the network code to not
run on the main thread. I think everything inside of the !prefToken
should be put in the AsyncTask but I'm not sure at this point. The
LinkedIn interface has been updated to show a certain user, Matthew
Lavin. However, the only thing that is returned is the first and last
name. I think this is a permissions problem that has to do with Linkedin
API and what the privacy level is by the users we are searching for. The
contacts that we search for are displayed in the same way as Ryan has
done the HistoryActivity.